adapters/reflex: Add OnComplete streamer
